requirements.txt
Pillow
py65
